feat: remove initial comment file
